+++ lagom/environment.py
@@ -23,8 +23,10 @@ import os
 from abc import ABC
 from typing import ClassVar, Optional
 
+from .exceptions import MissingEnvVariable, InvalidEnvironmentVariables
+
 try:
-    from pydantic.main import BaseModel
+    from pydantic.main import BaseModel, ValidationError
 except ImportError as error:
     raise ImportError(
         "Using Env requires pydantic to be installed. Try `pip install lagom[env]`"
@@ -48,15 +50,44 @@ class Env(ABC, BaseModel):
 
         :param kwargs:
         """
-        if len(kwargs) == 0:
-            prefix = f"{self.PREFIX}_" if self.PREFIX else ""
-            envs = os.environ.items()
-            super().__init__(
-                **{
-                    key.replace(prefix, "").lower(): value
-                    for (key, value) in envs
-                    if key.startswith(prefix)
-                }
-            )
-        else:
-            super().__init__(**kwargs)
+        try:
+            if len(kwargs) == 0:
+                prefix = self._prefix()
+                envs = os.environ.items()
+                super().__init__(
+                    **{
+                        key.replace(prefix, "").lower(): value
+                        for (key, value) in envs
+                        if key.startswith(prefix)
+                    }
+                )
+            else:
+                super().__init__(**kwargs)
+        except ValidationError as validation_error:
+            missing_field_errors = [
+                e
+                for e in validation_error.errors()
+                if e["type"] == "value_error.missing"
+            ]
+            if missing_field_errors:
+                env_names = self._env_names_from_pydantic_errors(missing_field_errors)
+                raise MissingEnvVariable(env_names) from validation_error
+            other_field_errors = [
+                e
+                for e in validation_error.errors()
+                if e["type"] != "value_error.missing"
+            ]
+            env_names = self._env_names_from_pydantic_errors(other_field_errors)
+            raise InvalidEnvironmentVariables(
+                env_names, str(validation_error)
+            ) from validation_error
+
+    def _env_names_from_pydantic_errors(self, missing_field_errors):
+        return [
+            f"{self._prefix()}{pyd_error['loc'][0]}".upper()
+            for pyd_error in missing_field_errors
+        ]
+
+    def _prefix(self):
+        prefix = f"{self.PREFIX}_" if self.PREFIX else ""
+        return prefix